fix(mbbs): rtkick stops truncating 32-bit delays through a 16-bit sign test
Task 17 of the abi-border plan. rtkick's negative-delay check tested bit
0x8000 of the raw argument and then cast the delay `as u16`. Correct for
Wg16 (int is two bytes), but under Wg32 (int is four bytes) the sign test
checks the wrong bit entirely and the cast discards the top two bytes:
rtkick(86400) -- one day, which LunatiX ships daily events for -- has low
16 bits 0x5180, whose bit 15 is clear, so it passed the "negative" check
and silently became 20864 seconds.

The sign test now anchors on A::INT_WIDTH * 8 - 1 (bit 15 for Wg16, bit 31
for Wg32) and the delay is carried whole as u32 end to end: Kick::delay,
Ended::Waiting/Bound's next_kick, and Wait::Until all widen from u16/
Option<u16> to u32/Option<u32>. mbbs-server/src/host.rs needed no change --
`secs.into()` already widens into Duration::from_secs's u64 regardless of
whether secs is u16 or u32.

crates/mbbs/tests/wg32_rtkick.rs adds the Wg32-side proof, through a real
PE import bound to rtkick and Host::run (rtkick itself is private, so this
is the only way to reach it from outside the crate): a negative-delay test
(0x8000_0000, negative as a 32-bit int but with bit 15 clear -- the victim
the old 0x8000 test cannot catch) written before the truncation test, per
the plan's own ordering.

+++ crates/mbbs/src/lib.rs
@@ -286,11 +286,11 @@ pub enum Ended<A: Abi> {
     /// other direction: `false` means the module ran out of work before the
     /// budget ran out, which is the only cheap evidence that the budget is
     /// more than enough.
-    Waiting { next_kick: u16, polls_cut: bool },
+    Waiting { next_kick: u32, polls_cut: bool },
 
     /// `max` passes were made and there is still work queued. A driver calls
     /// straight back.
-    Bound { next_kick: Option<u16> },
+    Bound { next_kick: Option<u32> },
 
     /// The module stopped, on the pass it stopped on.
     ///
@@ -357,7 +357,7 @@ pub enum Wait {
     /// Block until the transport delivers something.
     Blocked,
     /// Sleep at most this many whole seconds, waking early on input.
-    Until(u16),
+    Until(u32),
     /// Call `cycle` again now.
     Now,
     /// The module stopped. Shut the host down.
